[MIG] report_label [IMP] allow to create report.paperformat.label without creating each time report.paperformat item [FIX] Remove useless data that are specific Label: Agipa 114016 [REF] simplify UI, removing custom entry for ir.actions.server [ADD] migration script to migrate name field [IMP] Add label_background_color to allow to easily define background-color for label. (That can be usefull to debug label positions) [DOC] update screenshots to new V16 versions [IMP] Set body and html margin to 0 to to able to be predictive when designing a label sheet [IMP] replace style by class in the label template, reducing the size of the html code generated. [IMP] replace label_template by label_template_view_id on the ir.actions.server model, removing useless xml and python code. (provide migration scripts) [REF] Split wizard file into wizard and wizard line file, following OCA guidelines
