+++ report_label/__manifest__.py
@@ -1,7 +1,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 {
     "name": "Report Labels",
-    "version": "14.0.1.0.0",
+    "version": "16.0.1.0.0",
     "summary": "Print configurable self-adhesive labels reports",
     "author": "Iván Todorovich, Moka Tourisme, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/reporting-engine",
@@ -13,7 +13,6 @@
     ],
     "data": [
         "security/ir.model.access.csv",
-        "data/paperformat_label.xml",
         "views/ir_actions_server.xml",
         "views/report_paperformat_label.xml",
         "reports/report_label.xml",
