+++ report_label/models/ir_actions_server.py
@@ -9,8 +9,10 @@ class IrActionsServer(models.Model):
         selection_add=[("report_label", "Print self-adhesive labels")],
         ondelete={"report_label": "cascade"},
     )
-    label_template = fields.Char(
-        "Label QWeb Template",
+    label_template_view_id = fields.Many2one(
+        string="Label QWeb Template",
+        comodel_name="ir.ui.view",
+        domain=[("type", "=", "qweb")],
         help="The QWeb template key to render the labels",
         states={"report_label": [("required", True)]},
     )
@@ -20,27 +22,12 @@ class IrActionsServer(models.Model):
         states={"report_label": [("required", True)]},
     )
 
-    def report_label_associated_view(self):
-        """View the associated qweb templates"""
-        self.ensure_one()
-        action = self.env.ref("base.action_ui_view", raise_if_not_found=False)
-        if not action or len(self.label_template.split(".")) < 2:
-            return False
-        res = action.read()[0]
-        res["domain"] = [
-            ("type", "=", "qweb"),
-            "|",
-            ("name", "ilike", self.label_template.split(".")[1]),
-            ("key", "=", self.label_template),
-        ]
-        return res
-
     def _run_action_report_label_multi(self, eval_context=None):
         """Show report label wizard"""
         context = dict(self.env.context)
         context.update(
             {
-                "label_template": self.label_template,
+                "label_template_view_id": self.label_template_view_id.id,
                 "label_paperformat_id": self.label_paperformat_id.id,
                 "res_model_id": self.model_id.id,
             }
