+++ report_label/models/report_paperformat_label.py
@@ -3,8 +3,9 @@ from odoo import fields, models
 
 class ReportPaperformatLabel(models.Model):
     _name = "report.paperformat.label"
-    _inherits = {"report.paperformat": "paperformat_id"}
     _description = "Label Paper Format"
+
+    name = fields.Char(required=True)
 
     paperformat_id = fields.Many2one(
         "report.paperformat",
@@ -22,6 +23,7 @@ class ReportPaperformatLabel(models.Model):
         default=42.3,
         required=True,
     )
+    label_background_color = fields.Char(default="#FFFFFF")
     label_padding_top = fields.Float("Label Padding Top (mm)", default=2)
     label_padding_right = fields.Float("Label Padding Right (mm)", default=2)
     label_padding_bottom = fields.Float("Label Padding Bottom (mm)", default=2)
@@ -30,9 +32,3 @@ class ReportPaperformatLabel(models.Model):
     label_margin_right = fields.Float("Label Margin Right (mm)", default=2)
     label_margin_bottom = fields.Float("Label Margin Bottom (mm)", default=2)
     label_margin_left = fields.Float("Label Margin Left (mm)", default=2)
-
-    # Overload inherits defaults
-    orientation = fields.Selection(inherited=True, default="Portrait")
-    header_spacing = fields.Integer(inherited=True, default=0)
-    margin_top = fields.Float(inherited=True, default=7)
-    margin_bottom = fields.Float(inherited=True, default=7)
