+++ report_label/readme/USAGE.rst
@@ -1,5 +1,11 @@
-1. In the target model's tree view, select the records to print.
-2. Click *Action* and your label report action name.
-3. Select the number of labels per record to print, and click Print.
+* In the target model's tree view, select the records to print.
+* Click *Action* and your label report action name.
+* Select the number of labels per record to print, and click Print.
 
-.. image:: ../static/description/label_wizard.png
+you can optionaly define an offset
+
+.. image:: ../static/description/report_label_wizard_form.png
+
+The items will be printed in a pdf document.
+
+.. image:: ../static/description/report_result.png
