+++ report_label/reports/report_label.xml
@@ -2,33 +2,41 @@
 <odoo>
 
     <template id="report_label_template">
-        <t t-call="web.basic_layout">
-            <t t-set="full_width" t-value="True" />
-            <t t-set="label_style">
-                height:         <t t-esc="label_format['label_height']" />mm;
-                width:          <t t-esc="label_format['label_width']" />mm;
-                padding-top:    <t t-esc="label_format['label_padding_top']" />mm;
-                padding-right:  <t t-esc="label_format['label_padding_right']" />mm;
-                padding-bottom: <t t-esc="label_format['label_padding_bottom']" />mm;
-                padding-left:   <t t-esc="label_format['label_padding_left']" />mm;
-                margin-top:     <t t-esc="label_format['label_margin_top']" />mm;
-                margin-right:   <t t-esc="label_format['label_margin_right']" />mm;
-                margin-bottom:  <t t-esc="label_format['label_margin_bottom']" />mm;
-                margin-left:    <t t-esc="label_format['label_margin_left']" />mm;
-                display: inline-block;
-                overflow: hidden;
-                float: left;
-                position: relative;
-                page-break-inside: avoid;
-                box-sizing: border-box;
-            </t>
+
+        <t t-call="web.report_layout">
+            <style type="text/css">
+html, body {
+    margin: 0px;
+    padding: 0px;
+}
+.label_style {
+    height:             <t t-esc="label_format['label_height']" />mm;
+    width:              <t t-esc="label_format['label_width']" />mm;
+    padding-top:        <t t-esc="label_format['label_padding_top']" />mm;
+    padding-right:      <t t-esc="label_format['label_padding_right']" />mm;
+    padding-bottom:     <t t-esc="label_format['label_padding_bottom']" />mm;
+    padding-left:       <t t-esc="label_format['label_padding_left']" />mm;
+    margin-top:         <t t-esc="label_format['label_margin_top']" />mm;
+    margin-right:       <t t-esc="label_format['label_margin_right']" />mm;
+    margin-bottom:      <t t-esc="label_format['label_margin_bottom']" />mm;
+    margin-left:        <t t-esc="label_format['label_margin_left']" />mm;
+    background-color:   <t t-esc="label_format['label_background_color']" />;
+    display: inline-block;
+    overflow: hidden;
+    float: left;
+    position: relative;
+    page-break-inside: avoid;
+    box-sizing: border-box;
+}
+            </style>
+
             <!-- Offset: Skip the first [offset] labels -->
             <t t-foreach="range(0, offset)" t-as="i">
-                <div t-att-style="label_style" />
+                <div class="label_style" />
             </t>
             <t t-foreach="lines" t-as="line">
                 <t t-foreach="range(0, line['quantity'])" t-as="i">
-                    <div t-att-style="label_style">
+                    <div class="label_style">
                         <t t-call="{{label_template}}">
                             <t
                                 t-set="record"
