+++ report_label/views/ir_actions_server.xml
@@ -5,42 +5,17 @@
         <field name="model">ir.actions.server</field>
         <field name="inherit_id" ref="base.view_server_action_form" />
         <field name="arch" type="xml">
-            <header position="inside">
-                <button
-                    name="report_label_associated_view"
-                    type="object"
-                    string="View QWeb templates"
-                    attrs="{'invisible':['|', ('state', '!=', 'report_label'), ('label_template', '=', False)]}"
-                    help="Display a button in the sidebar of related model to open a wizard"
-                />
-            </header>
             <field name="type" position="after">
                 <field
                     name="label_paperformat_id"
                     attrs="{'invisible': [('state', '!=', 'report_label')]}"
                 />
                 <field
-                    name="label_template"
+                    name="label_template_view_id"
                     attrs="{'invisible': [('state', '!=', 'report_label')]}"
                 />
             </field>
         </field>
     </record>
-
-    <record id="report_label_action" model="ir.actions.act_window">
-        <field name="name">Label Reports</field>
-        <field name="res_model">ir.actions.server</field>
-        <field name="view_mode">tree,form</field>
-        <field name="domain">[("state", "=", "report_label")]</field>
-        <field name="context">{"default_state": "report_label"}</field>
-    </record>
-
-    <menuitem
-        id="report_label_menu"
-        name="Label Reports"
-        action="report_label_action"
-        parent="base.reporting_menuitem"
-        sequence="3"
-    />
 
 </odoo>
