+++ report_label/wizards/__init__.py
@@ -1,1 +1,2 @@
 from . import report_label_wizard
+from . import report_label_wizard_line
